move sync test to consensus_test
